rename: js파일명 camelCase로 변경
