+++ components/footer/footerComponent.js
