+++ components/gnb/gnbComponent.js
