+++ components/star/starComponent.js
