+++ index.html
@@ -130,7 +130,7 @@
       </article>
     </section>
     <footer-component></footer-component>
-    <script type="module" src="/components/gnb/gnb-component.js"></script>
-    <script type="module" src="/components/footer/footer-component.js"></script>
+    <script type="module" src="/components/gnb/gnbComponent.js"></script>
+    <script type="module" src="/components/footer/footerComponent.js"></script>
   </body>
 </html>
