+++ shared/index.html
@@ -24,16 +24,13 @@
     </main>
     <footer-component></footer-component>
 
-    <script type="module" src="/components/gnb/gnb-component.js"></script>
+    <script type="module" src="/components/gnb/gnbComponent.js"></script>
     <script
       type="module"
       src="/components/searchBar/searchBarComponent.js"
     ></script>
-    <script
-      type="module"
-      src="/components/card/card-list-component.js"
-    ></script>
-    <script type="module" src="/components/footer/footer-component.js"></script>
+    <script type="module" src="/components/card/cardListComponent.js"></script>
+    <script type="module" src="/components/footer/footerComponent.js"></script>
     <script type="module" src="/components/folder-info/folderInfo.js"></script>
   </body>
 </html>
